apps/api/pyproject.toml
[project]
name = "healthtwin-api"
version = "0.1.0"
description = "HealthTwin healthcare-native backend API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "openai>=1.0.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
